move counter to monitor
