+++ src/bin/event_orderbook.rs
@@ -8,7 +8,8 @@ use kucoin_arbitrage::broker::orderchange::kucoin::task_pub_orderchange_event;
 use kucoin_arbitrage::broker::symbol::filter::symbol_with_quotes;
 use kucoin_arbitrage::broker::symbol::kucoin::{format_subscription_list, get_symbols};
 use kucoin_arbitrage::event::{orderbook::OrderbookEvent, orderchange::OrderChangeEvent};
-use kucoin_arbitrage::model::{counter::Counter, orderbook::FullOrderbook};
+use kucoin_arbitrage::model::orderbook::FullOrderbook;
+use kucoin_arbitrage::monitor::counter::Counter;
 use kucoin_arbitrage::monitor::task::task_log_mps;
 use std::sync::Arc;
 use tokio::signal::unix::{signal, SignalKind};
