+++ src/bin/test_kucoin_order_private_latency.rs
@@ -7,7 +7,7 @@ use kucoin_arbitrage::broker::orderchange::kucoin::task_pub_orderchange_event;
 use kucoin_arbitrage::broker::symbol::kucoin::{format_subscription_list, get_symbols};
 use kucoin_arbitrage::event::order::OrderEvent;
 use kucoin_arbitrage::event::orderchange::OrderChangeEvent;
-use kucoin_arbitrage::model::counter::Counter;
+use kucoin_arbitrage::monitor::counter::Counter;
 use kucoin_arbitrage::model::order::{LimitOrder, OrderSide, OrderType};
 use kucoin_arbitrage::strings::generate_uid;
 use kucoin_arbitrage::{broker::symbol::filter::symbol_with_quotes, monitor};
